projection_app/requirements.txt
numpy==2.4.2
PyOpenGL==3.1.10
PyQt6==6.10.2
PyQt6-Qt6==6.10.2
PyQt6_sip==13.11.0
